Run Linter on file save & util to set cwd
